src/database/oracle_connection.py
"""
Oracle AI Vector DB Connection Module

This module handles connection to Oracle AI Vector Database.
Note: This is a placeholder for future Oracle AI Vector DB integration.
"""

import os
from typing import Optional
from dotenv import load_dotenv

load_dotenv()


class OracleConnection:
    """Manages Oracle AI Vector DB connection and operations."""
    
    def __init__(self):
        """Initialize Oracle connection with environment variables."""
        self.user = os.getenv("ORACLE_USER", "")
        self.password = os.getenv("ORACLE_PASSWORD", "")
        self.dsn = os.getenv("ORACLE_DSN", "")
        self.connection: Optional[object] = None
    
    def connect(self) -> bool:
        """Establish connection to Oracle AI Vector DB."""
        try:
            import oracledb
            
            self.connection = oracledb.connect(
                user=self.user,
                password=self.password,
                dsn=self.dsn
            )
            print(f"✓ Successfully connected to Oracle AI Vector DB")
            return True
        except ImportError:
            print("✗ oracledb package not installed. Install with: pip install oracledb")
            return False
        except Exception as e:
            print(f"✗ Failed to connect to Oracle AI Vector DB: {e}")
            return False
    
    def close(self):
        """Close the Oracle connection."""
        if self.connection:
            self.connection.close()
            print("✓ Oracle connection closed")
    
    def test_connection(self):
        """Test the database connection."""
        if not self.connection:
            print("✗ Connection not established")
            return False
        
        try:
            cursor = self.connection.cursor()
            cursor.execute("SELECT 1 FROM DUAL")
            result = cursor.fetchone()
            cursor.close()
            
            if result:
                print("✓ Oracle AI Vector DB connection test successful")
                return True
            return False
        except Exception as e:
            print(f"✗ Connection test failed: {e}")
            return False


def main():
    """Test the Oracle connection."""
    print("=" * 50)
    print("Oracle AI Vector DB Connection Test")
    print("=" * 50)
    
    conn = OracleConnection()
    
    if conn.connect():
        conn.test_connection()
        conn.close()
    else:
        print("\nPlease check your .env file and ensure Oracle AI Vector DB credentials are correct.")


if __name__ == "__main__":
    main()

