requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# Neo4j Driver
neo4j>=5.15.0

# LangChain and Related
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-neo4j>=0.0.1

# Oracle AI Vector DB
oracledb>=2.0.0

# LLM Integration (Ollama support)
langchain-ollama>=0.0.1

# Utilities
pyyaml>=6.0.1

